02.6 registrar e excluir chaves no bcb

+++ src/main/kotlin/br/com/zup/pix/remove/RemoveChavePixService.kt
@@ -2,7 +2,11 @@ package br.com.zup.pix.remove
 
 import br.com.zup.pix.ChavePixRepository
 import br.com.zup.pix.exceptions.ChavePixNaoEncontradaException
+import br.com.zup.pix.externo.BcbClient
+import br.com.zup.pix.externo.DeletePixKeyRequest
+import br.com.zup.pix.model.ContaAssociada
 import br.com.zup.pix.validation.ValidUUID
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import java.util.*
 import javax.inject.Inject
@@ -12,7 +16,8 @@ import javax.validation.constraints.NotBlank
 
 @Validated
 @Singleton
-class RemoveChavePixService(@Inject val repository : ChavePixRepository) {
+class RemoveChavePixService(@Inject val repository : ChavePixRepository,
+                            @Inject val bcbClient: BcbClient) {
 
     @Transactional
     fun remove(
@@ -24,6 +29,16 @@ class RemoveChavePixService(@Inject val repository : ChavePixRepository) {
 
         val chave = repository.findByIdAndClienteId(uuidPixId, uuidClienteId)
             .orElseThrow{ ChavePixNaoEncontradaException("Chave pix nao encontrado ou nao associada a um cliente") }
+
+        //se encontrarmos a chave primeiro tentamos excluir no BCB e só depois na nossa aplicacao
+        val respostaBcb = bcbClient.deletaChaveBcb(
+            keyPath = chave.chave,
+            request = DeletePixKeyRequest(chave.chave, ContaAssociada.ISPB)
+        )
+
+        if (respostaBcb.status != HttpStatus.OK) {
+            throw IllegalStateException("Erro ao remover chave Pix no Banco Central do Brasil (BCB)")
+        }
 
         repository.deleteById(uuidPixId)
     }
